Rework the way we store lecture information.
Base unit is now a post rather than a lecture.

There is a ParentPost and SubPost objects, should be obvious
how they work.

New Post class, which it subclassed by ParentPost and Subpost
models.

New Materials model for holding information related to materials
for use by post objects.

Updated the template to and css to use this new layout.