Add Subject, AcademicYear and History models.
Add more hierarchy with the addition of Subject and AcademicYear to provide
 finer grain filtering for later.

Add the History and HistoryMixIn to provide a method of storing posts from the
past. This way users can look at material provided in previous years.

Update the view so that I now filters out the correct posts, rather than relying
on the template. This will make it easier to show posts for a specific time
frame easier.

Update the template to match the view changes.