Add the files layer in.
We now have a file management layer. This works using documents and blobs.

Blobs are actuall files stored on the cloud, they are used to deduplicate as two blobs with the same md5 hash in the same container should never happen.

The docs are what everything interacts with, these hide the fact the muliple uploads may actually only result in a single blob.

Derived documents are files that are derived from uploaded content. For example preview pngs or pdfs. These are tightly coupled to the blob they are derived from, thus preventing us from wasting CPU power converting the same file multiple times.

Between derived documents and document 90% of the deduplication layer is hidden.

Finally there are some helper functions to aid in the creating of documents and derived documents. Primarly these functions handle deduplication and blob management.