Massive improvement to file management and deduplication
Now you just need to interact with the `document` and `deriveddocument` objects, by assigning file to their `file` attributes they will automatically find or create the necessary blob.

Deleting a `document` or `deriveddocument` will also result in the underlying blob being deleted if there are no longer and documents pointing to it.