Override the delete() method of blob files.
The blob object now takes every opportunity to override the delete() method on its files to prevent people deleting the files that underpins a blob.

Unfortunately there are a couple of issues with new blob objects that have never been saved as there appears to be no hook that will allow me to reliably override the delete() method. (This also applies to blob's created using objects.create, but only on the returned instance, if the instance is retrieved from the DB or saved after using create() all is fine.)

As a result I have also updated the test_helpers to make sure that save() is called on returned blob instances.