Add quite a sophisticated admin interface for Document's
Added an admin interface which plays nice with the way documents and blobs are meant to be created. The admin site has been modified to hide the use of blobs in the document page. Instead there is just a file upload widget.

When a new document is created or a new file is uploaded to a document it is passed through the standard de-duplication layer and blob are created and destroyed as necessary.