Minor change to admin pages.
We no longer hide the `file_name` field in an add page. The code is still there but commented out.

I have done this because `file_name` like `title` is not required in the form and is automatically generated later from the uploaded file if the fields are left blank.