Add the start of q and a
Started working on the q and a part of the site. This mean adding the app to settings and also adding the necessary urls.

Also fixed up the reputation app a little bit better. This includes getting rid of the `UserReputation` model and just adding a reputation field to the `UserProfile` model instead.

Also added the template that we need and the css.