Move jQuery hook for csrf safe ajax into separate js and html files
The jQuery hook to pass back the csrf cookie with ajax requests in now in a separate file which can be included by including `templates/csrf_js.html` in the head block of a template.

Additionally a footer block and jQuery script has been added to `base.html` to make my life easier when using the above.