Huge overhaul of look and feel.
Reworked the layout of the site to use a basic grid layout. So far I have updated most of the modules templates to use this new layout.

Also I have refactored all of the SCSS to make it easier to maintain with different SCSS sections in different files that are imported into the final file.