Add file type priorities and display strings.
Added to the settings is the priorities of the different file types, this will effect the order in which they are listed in downloads. This only effects order between files derived from the same source.

Also add display strings which are presented to the user instead of the file type.

Finally updated the file app to make use of these new additions with new helper functions and properties.

Note:
The way I calculate the order to return derived files is debatably correct, and could use some more work in the future.