Create a new download app to providing downloads.
The new download app provides a couple of urls for downloading Documents and DerivedDocuments.

It takes the request and queries amazon S3 to generate a temp url to the file requested, then redirects the user the the S3 url for download.

Additionally it modifies the content-disposition header to ensure that filename the end user sees matches the file name in the document object.