Expand the url property to generate download links.
The url property now generates valid urls for the download app, thus allowing people to download files.

Additionally when the versions property on the Document model is used, the resulting DerivedDocuments have an additional "original_document" attribute set, point back to current Document instance.

This is used when generating the download url's to inform the download app which document the deriveddocument is related too, and thus adjust the filenames to be more appropriate.