Move colour definitions and make title and module look better.
Colour definitions where moved out into an importable file so that they can be imported into separate CCS files, and provide us with a single point to modify them from.

The titles are now button link, that border on them makes slightly 3D and appear to come out of the page.