Add vote up and down buttons.
Add the html and css to add the vote up and down buttons. There is currently no JS linking them to the backend.

To add them I had to use a table to create two columns, with the first column for the buttons and the right for the post content. The entire table has it width limited to the width of the container by CSS.

The left column is fixed width, and the right fills the remaining space.