Implement question and answer voting.
In question.html there is JS to do ajax voting so the page does not reload when a person votes.

This JS posts to a 'vote' url defined in urls.py which in turn calls a voting view in views.py.

This view uses the functions defined by the reputation mixin to increment and decrement the objects vote score.

The reputation mixin and the reputation model has both be upgraded to provide some more useful methods to allow me to clear votes, and check to see which way a person has voted.

Also the question view has been upgraded to get information on how a user has voted, and annotate the question and answer objects with this information so the template can use it. The view should also have received a small speed boost due to the usage of select_related() to reduce the number of db hits on each page load.

Finally the post.html template uses the object annotation to set the 'voted' css class on the appropriate vote button. The JS uses this value to determine whether a vote is being cleared or not.