Replace pagedown with form and widget.
This is quite a big change. I have reworked the way the pagedown editor is used. It is now a widget that can be included in a form, this way I can use django's form framework to handle use responses instead of trying to write my own stuff.

This had meant that I have got rid of pagedown.html, and instead replaced it with the pagedown app which contains the necessary html in a widget.

The questions template and view has also been updated to reflect this, now I have a very simple form for the Post Answers section, which uses the Pagedown widget.

In the view I generate a form instance, and pass it to the templates Context. In the template its self I just use {{ answer_form.as_p }} in inject the pagedown html, and {{ answer_form.media }} to inject the html and css.

Unfortunately this may break django-compressor, but that is a problem for a later date.