Handle voting when person is logged out.
No nothing happens, instead of a 404 error and redirect to a login page, which of cause just caused javascript errors.

Now I just return a json lump with a error in it. This make client JS do nothing, with no JS exception or any visual change. Really I should inform the user, but that will come later.