Add answer POST view.
New url in the q and a urls for the new view which will process answer posts from the site.

Currently the view only verifies the answer and adds it to the DB. Currently it does not redirect the user or anything like that. I may modify how the POST happens at a later date.