Move tags to ParentPost and SubPost.
As a bit of a hack to get around a small bug in Django ORM or contenttype app I have had to take the tags attribute of the Post supermodel and onto the subclasses.

This is because Django taggit adds a GenericRelation attribute to the model to for various reasons, one of which is to have the related TaggedItem recored deleted when the model it's related to is deleted.

This functionally does not seem to work in Django version 1.4. See Django ticket #19149.