Improve home page column handling and add "Recent Questions"
Improved the way multiple columns are handled. The number of items in each column is dynamically determined based on the number of items that we are trying to display.

From the dictionary of all the "blobs" that need to be displayed the view will use an iterator to place half of the blobs in one column and the rest in the other. Using the iterator allows the blobs to be laid out from left-to-right, then continuing onto the row below rather than stacking the first of the items in the first column and the second half in the second column.