Create PDF's on blob save rather than Document save
I have changed the way the create_pdf task is both fired, and how it works. The task is now fired on the save of a blob object rather that a document object.

I do this so that if the users changes the file in a Document object the new file (because it will have probably have created a new blob) will be converted as expected, rather than having to create silly logic to deal with this case from the document perspective.