Only fire create_pdf on creation, and check for PDF before save.
Once again we only fire the create_pdf task on the creation of a ParentBlob.

In addition we also do one final check for a derived file of type PDF just before we save the new derived_document.

All of this is done to reduce the chance of create_pdf being fired twice or more at the same time on the same blob, and having all but one fail when it comes to saving the derived file.

Before this the race condition did not causes any real problems, just an exception and some wasted CPU cycles, now the race condition is probably eliminated, and if it hasn't a far stricter set of conditions must be met before an exception will be raised.