Add years to front page for anonymous users.
Adding each year and all of it modules to the home page of the site if being accessed by a anonymous user. (I may change this in the future so that they appear for logged in users as well.)

To do this I had to use a bodge on the "items" dictionary in the view, and also the template.

Also I had to add a related name for modules to the AcademicYear model to make my life easier.