Improve extraction error handleing.
I created a new field on blobs to include an extraction_error field. This way when it has an issue trying to extract the content of a document it can mark the blob.

By doing this I can reduce the number of emails I get. An error is only raised the first time we encouter an index problem, and not on subsiquent updates.

If the extraction state changes then the flag will be auto unset. It is not modifiable from the admin site.

Finally I have stopped addressing the blob directly in the search index. I have created a set of properties on the documents to hide the blobs, along with a flag to monitor when a blob is modified so that it is saved when the doc is saved. (This way you never need to know if you interacting with the blob or not, the doc will worry about it. This is the way it should have been from the start.)

P.S. There is a migration in here too.