Reverse the way lecturer and student profiles link with userprofiles.
Before the UserProfile used a nullable one-to-one to point at lecturer and student profiles. Now the lecturer and student profiles uses a NON-nullable one-to-one to point at the UserProfile. Doing this means that there can never be orphaned student and lecturer profiles, and that they get cascade deleted with UserProfiles.

I have also updated the admin interface, and the unicode representation of student and lecture profiles.

This is how it should have been done from the start. I have updated the is_student and is_lecture methods to use a DoesNotExistError exception and try...except to check if the user is a student of lecrurer.

Finally I have created the necessary schema and data migrations to go back and forth.