Create tasks to creating PDF's and PNG's.
This is the first basic version. It relies on a custom version of unoconv to work because the stock version does not react well to SIGTERM.

The PDF convertion uses the above unoconv to control LibreOffice and generate the PDF files which are then uploaded to S3.

The PNG creation is taken straight from my old site, but modified to assume that it can talk directly to the DB and S3. This makes everything that much easier to work with.

There is a special queue just for conversion tasks so that the workers can be setup with a special concurrancy to deal with the long run times of these tasks.

Finally add a setting to ensure that boto will never overwrite when a new one gets uploaded with the same name but will do what every other djagno storage backend does and finds a new name for the file.