Move "file_type" field from blob to documents.
I have moved it because it makes more sense on the various document objects. This way querys are easier to write, and in real use people should never be interacting with blobs, so this change is transparent to them.

The primary motivator for this is being able to have a unique constain on derived documents accross the index, derived_from, and file_type fields. This allows be to use the index for its original purpose. Providing an order to derived files of the same type (e.g. PNG's).