Add sorl_thumbnail to generate preview images.
I am using sorl_thumbnail to generate thumbnails. The results are cached in S3, and a key-value store. I have added the app to settings, and updated the materials template to use it. This need further work, but has shown that sorl_thumbnail is the best package for the job (nice and easy to use).